Recommender/requirements.txt
Flask
flask-cors
pydantic
arabic_reshaper
python-bidi
scikit-learn
numpy
pymongo
